+++ src/main/scala/control/BaseChartCanvas.scala
@@ -1,0 +1,33 @@
+package bigvis
+package control
+
+import javafx.scene.canvas.{Canvas, GraphicsContext}
+import javafx.scene.paint.Color
+
+
+object RenderHelper {
+  def drawContrastLine(gc: GraphicsContext, background: Color,
+                       x1: Double, y1: Double, x2: Double, y2: Double): Unit = {
+    gc.save()
+    gc.setStroke(background)
+    gc.setLineWidth(gc.getLineWidth * 5)
+    gc.strokeLine(x1, y1, x2, y2)
+    gc.restore()
+    gc.strokeLine(x1, y1, x2, y2)
+  }
+
+  def drawContrastText(gc: GraphicsContext, background: Color, text: String,
+                       x: Double, y: Double): Unit = {
+    gc.save()
+    gc.setStroke(background)
+    gc.setLineWidth(5)
+    gc.strokeText(text, x, y)
+    gc.restore()
+    gc.fillText(text, x, y)
+  }
+}
+
+
+class BaseChartCanvas extends Canvas {
+  override def isResizable: Boolean = true
+}
